chore(client): refactor restart component
