+++ client/src/pages/app/settings.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import * as yup from 'yup';
 import { useHistory, useParams } from 'react-router-dom';
 import { toast } from 'react-toastify';
@@ -6,44 +6,20 @@ import { Header } from '../../modules/layout/Header';
 import {
   useAppByIdQuery,
   useDestroyAppMutation,
-  useRestartAppMutation,
-  useAppRestartLogsSubscription,
   DashboardDocument,
-  RealTimeLog,
 } from '../../generated/graphql';
 import { useFormik } from 'formik';
-import {
-  TabNav,
-  TabNavLink,
-  Button,
-  Modal,
-  ModalTitle,
-  ModalDescription,
-  ModalButton,
-  Terminal,
-  FormInput,
-  FormHelper,
-} from '../../ui';
+import { TabNav, TabNavLink, Button, FormInput, FormHelper } from '../../ui';
 import { AppProxyPorts } from '../../modules/appProxyPorts/AppProxyPorts';
+import { AppRestart } from '../../modules/app/AppRestart';
 
 export const Settings = () => {
   const { id: appId } = useParams<{ id: string }>();
-  const [isRestartAppModalOpen, setIsRestartAppModalOpen] = useState(false);
-  const [arrayOfRestartLogs, setArrayOfRestartLogs] = useState<RealTimeLog[]>(
-    []
-  );
-  const [isTerminalVisible, setIsTerminalVisible] = useState(false);
-  const [restartLoading, setRestartLoading] = useState(false);
-  const [processStatus, setProcessStatus] = useState<
-    'running' | 'notStarted' | 'finished'
-  >('notStarted');
   let history = useHistory();
   const [
     destroyAppMutation,
     { loading: destroyAppMutationLoading },
   ] = useDestroyAppMutation();
-
-  const [restartAppMutation] = useRestartAppMutation();
 
   const { data, loading /* error */, refetch } = useAppByIdQuery({
     variables: {
@@ -52,39 +28,6 @@ export const Settings = () => {
     ssr: false,
     skip: !appId,
   });
-
-  useAppRestartLogsSubscription({
-    onSubscriptionData: (data) => {
-      const logsExist = data.subscriptionData.data?.appRestartLogs;
-      if (logsExist) {
-        setArrayOfRestartLogs((currentLogs) => {
-          return [...currentLogs, logsExist];
-        });
-        if (
-          logsExist.type === 'end:success' ||
-          logsExist.type === 'end:failure'
-        ) {
-          setProcessStatus('finished');
-        }
-      }
-    },
-  });
-
-  const handleRestartApp = async () => {
-    try {
-      await restartAppMutation({
-        variables: {
-          input: {
-            appId,
-          },
-        },
-      });
-      setIsTerminalVisible(true);
-      setRestartLoading(true);
-    } catch (e) {
-      toast.error(e.message);
-    }
-  };
 
   const DeleteAppNameSchema = yup.object().shape({
     appName: yup
@@ -166,71 +109,8 @@ export const Settings = () => {
                 Update the settings of your app.
               </p>
             </div>
-
             <AppProxyPorts appId={app.id} />
-            <h1 className="text-md font-bold mt-6">Restart app</h1>
-            <p className="text-gray-400">
-              Restart your dokku app and see logs in real time.
-            </p>
-
-            <div className="mt-1">
-              <Button
-                onClick={() => setIsRestartAppModalOpen(true)}
-                color="grey"
-                className="mt-2"
-              >
-                Restart
-              </Button>
-            </div>
-            {isRestartAppModalOpen && (
-              <Modal>
-                <ModalTitle>Restart app</ModalTitle>
-                <ModalDescription>
-                  {isTerminalVisible ? (
-                    <React.Fragment>
-                      <p className="mb-2 ">Restarting {app.name}</p>
-                      <p className="text-gray-500 mb-2">
-                        Restarting the app usually takes couple of minutes.
-                        Breathe in, breathe out, logs are about to appear below:
-                      </p>
-                      <Terminal className={'w-6/6'}>
-                        {arrayOfRestartLogs.map((log) => (
-                          <p
-                            key={arrayOfRestartLogs.indexOf(log)}
-                            className="text-s leading-5"
-                          >
-                            {log.message}
-                          </p>
-                        ))}
-                      </Terminal>
-                    </React.Fragment>
-                  ) : (
-                    <p>
-                      {`Are you sure, you want to restart ${app.name} app ?`}
-                    </p>
-                  )}
-                </ModalDescription>
-                <ModalButton
-                  ctaFn={() => {
-                    setProcessStatus('running');
-                    handleRestartApp();
-                  }}
-                  ctaText={'Restart'}
-                  otherButtonText={'Cancel'}
-                  isCtaLoading={isTerminalVisible ? false : restartLoading}
-                  isCtaDisabled={isTerminalVisible}
-                  isOtherButtonDisabled={processStatus === 'running'}
-                  closeModal={() => {
-                    setIsRestartAppModalOpen(false);
-                    refetch({ appId });
-                    setRestartLoading(false);
-                    setIsTerminalVisible(false);
-                    setProcessStatus('notStarted');
-                  }}
-                />
-              </Modal>
-            )}
-
+            <AppRestart appId={app.id} />
             <h1 className="text-md font-bold py-5">Delete app</h1>
             <p className="text-gray-400">
               This action cannot be undone. This will permanently delete{' '}
